+++ mne/viz/raw.py
@@ -14,16 +14,18 @@ import numpy as np
 
 from ..externals.six import string_types
 from ..io.pick import (pick_types, _pick_data_channels, pick_info,
-                       _PICK_TYPES_KEYS)
+                       _PICK_TYPES_KEYS, pick_channels)
 from ..io.proj import setup_proj
 from ..utils import verbose, get_config
 from ..time_frequency import psd_welch
 from .topo import _plot_topo, _plot_timeseries, _plot_timeseries_unified
 from .utils import (_toggle_options, _toggle_proj, tight_layout,
-                    _layout_figure, _plot_raw_onkey, figure_nobar,
-                    _plot_raw_onscroll, _mouse_click, plt_show,
+                    _layout_figure, _plot_raw_onkey, figure_nobar, plt_show,
+                    _plot_raw_onscroll, _mouse_click, _find_channel_idx,
                     _helper_raw_resize, _select_bads, _onclick_help,
-                    _setup_browser_offsets, _compute_scalings)
+                    _setup_browser_offsets, _compute_scalings, plot_sensors,
+                    _radio_clicked, _set_radio_button, _handle_topomap_bads,
+                    _change_channel_group)
 from ..defaults import _handle_default
 from ..annotations import _onset_to_seconds
 
@@ -100,7 +102,8 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
     start : float
         Initial time to show (can be changed dynamically once plotted).
     n_channels : int
-        Number of channels to plot at once. Defaults to 20.
+        Number of channels to plot at once. Defaults to 20. Has no effect if
+        ``order`` is 'position' or 'selection'.
     bgcolor : color object
         Color of the background.
     color : dict | color object | None
@@ -129,10 +132,13 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
 
     remove_dc : bool
         If True remove DC component when plotting data.
-    order : 'type' | 'original' | array
+    order : 'type' | 'original' | 'selection' | 'position' | array of int
         Order in which to plot data. 'type' groups by channel type,
-        'original' plots in the order of ch_names, array gives the
-        indices to use in plotting.
+        'original' plots in the order of ch_names, 'selection' uses
+        Elekta's channel groupings (only works for Neuromag data) and
+        'positions' groups the channels by the positions of the sensors.
+        If array, the order is determined by the indices in the array.
+        Defaults to 'type'.
     show_options : bool
         If True, a dialog for options related to projection is shown.
     title : str | None
@@ -258,16 +264,14 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
     # put them back to original or modified order for natral plotting
     reord = np.argsort(inds)
     types = [types[ri] for ri in reord]
-    if isinstance(order, str):
+    if isinstance(order, string_types):
         if order == 'original':
             inds = inds[reord]
+        elif order in ['selection', 'position']:
+            selections, fig_selection = _setup_browser_selection(raw, order)
         elif order != 'type':
             raise ValueError('Unknown order type %s' % order)
-    elif isinstance(order, np.ndarray):
-        if not np.array_equal(np.sort(order),
-                              np.arange(len(info['ch_names']))):
-            raise ValueError('order, if array, must have integers from '
-                             '0 to n_channels - 1')
+    elif isinstance(order, (np.ndarray, list)):
         # put back to original order first, then use new order
         inds = inds[reord][order]
 
@@ -288,9 +292,16 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
     params = dict(raw=raw, ch_start=0, t_start=start, duration=duration,
                   info=info, projs=projs, remove_dc=remove_dc, ba=ba,
                   n_channels=n_channels, scalings=scalings, types=types,
-                  n_times=n_times, event_times=event_times,
+                  n_times=n_times, event_times=event_times, inds=inds,
                   event_nums=event_nums, clipping=clipping, fig_proj=None)
 
+    if isinstance(order, string_types) and order in ['selection', 'position']:
+        params['fig_selection'] = fig_selection
+        params['selections'] = selections
+        params['radio_clicked'] = partial(_radio_clicked, params=params)
+        fig_selection.radio.on_clicked(params['radio_clicked'])
+        pick_callback = partial(_region_picked, params=params)
+        fig_selection.canvas.mpl_connect('pick_event', pick_callback)
     _prepare_mne_browse_raw(params, title, bgcolor, color, bad_color, inds,
                             n_channels)
 
@@ -298,17 +309,16 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
     event_lines = [params['ax'].plot([np.nan], color=event_color[ev_num])[0]
                    for ev_num in sorted(event_color.keys())]
 
-    params['plot_fun'] = partial(_plot_raw_traces, params=params, inds=inds,
-                                 color=color, bad_color=bad_color,
-                                 event_lines=event_lines,
+    params['plot_fun'] = partial(_plot_raw_traces, params=params, color=color,
+                                 bad_color=bad_color, event_lines=event_lines,
                                  event_color=event_color)
 
     if raw.annotations is not None:
         segments = list()
         segment_colors = dict()
         # sort the segments by start time
-        order = raw.annotations.onset.argsort(axis=0)
-        descriptions = raw.annotations.description[order]
+        ann_order = raw.annotations.onset.argsort(axis=0)
+        descriptions = raw.annotations.description[ann_order]
         color_keys = set(descriptions)
         color_vals = np.linspace(0, 1, len(color_keys))
         for idx, key in enumerate(color_keys):
@@ -317,9 +327,9 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
             else:
                 segment_colors[key] = plt.cm.summer(color_vals[idx])
         params['segment_colors'] = segment_colors
-        for idx, onset in enumerate(raw.annotations.onset[order]):
+        for idx, onset in enumerate(raw.annotations.onset[ann_order]):
             annot_start = _onset_to_seconds(raw, onset)
-            annot_end = annot_start + raw.annotations.duration[order][idx]
+            annot_end = annot_start + raw.annotations.duration[ann_order][idx]
             segments.append([annot_start, annot_end])
             ylim = params['ax_hscroll'].get_ylim()
             dscr = descriptions[idx]
@@ -368,6 +378,19 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
     # deal with projectors
     if show_options is True:
         _toggle_options(None, params)
+    # initialize the first selection set
+    if order in ['selection', 'position']:
+        _radio_clicked(fig_selection.radio.labels[0]._text, params)
+        callback_selection_key = partial(_selection_key_press, params=params)
+        callback_selection_scroll = partial(_selection_scroll, params=params)
+        callback_close = partial(_close_event, params=params)
+        params['fig'].canvas.mpl_connect('close_event', callback_close)
+        params['fig_selection'].canvas.mpl_connect('close_event',
+                                                   callback_close)
+        params['fig_selection'].canvas.mpl_connect('key_press_event',
+                                                   callback_selection_key)
+        params['fig_selection'].canvas.mpl_connect('scroll_event',
+                                                   callback_selection_scroll)
 
     try:
         plt_show(show, block=block)
@@ -375,6 +398,31 @@ def plot_raw(raw, events=None, duration=10.0, start=0.0, n_channels=20,
         plt_show(show)
 
     return params['fig']
+
+
+def _selection_scroll(event, params):
+    """Callback for scroll in selection dialog."""
+    if event.step < 0:
+        _change_channel_group(-1, params)
+    elif event.step > 0:
+        _change_channel_group(1, params)
+
+
+def _selection_key_press(event, params):
+    """Callback for keys in selection dialog."""
+    if event.key == 'down':
+        _change_channel_group(-1, params)
+    elif event.key == 'up':
+        _change_channel_group(1, params)
+    elif event.key == 'escape':
+        _close_event(event, params)
+
+
+def _close_event(event, params):
+    """Callback for closing of raw browser with selections."""
+    import matplotlib.pyplot as plt
+    plt.close(params['fig_selection'])
+    plt.close(params['fig'])
 
 
 def _label_clicked(pos, params):
@@ -385,17 +433,23 @@ def _label_clicked(pos, params):
     text = labels[line_idx].get_text()
     if len(text) == 0:
         return
-    ch_idx = params['ch_start'] + line_idx
+    if 'fig_selection' in params:
+        ch_idx = _find_channel_idx(text, params)
+        _handle_topomap_bads(text, params)
+    else:
+        ch_idx = [params['ch_start'] + line_idx]
     bads = params['info']['bads']
     if text in bads:
         while text in bads:  # to make sure duplicates are removed
             bads.remove(text)
         color = vars(params['lines'][line_idx])['def_color']
-        params['ax_vscroll'].patches[ch_idx].set_color(color)
+        for idx in ch_idx:
+            params['ax_vscroll'].patches[idx].set_color(color)
     else:
         bads.append(text)
         color = params['bad_color']
-        params['ax_vscroll'].patches[ch_idx].set_color(color)
+        for idx in ch_idx:
+            params['ax_vscroll'].patches[idx].set_color(color)
     params['raw'].info['bads'] = bads
     _plot_update_raw_proj(params, None)
 
@@ -582,18 +636,40 @@ def _prepare_mne_browse_raw(params, title, bgcolor, color, bad_color, inds,
 
     # populate vertical and horizontal scrollbars
     info = params['info']
-    for ci in range(len(info['ch_names'])):
-        this_color = (bad_color if info['ch_names'][inds[ci]] in info['bads']
-                      else color)
-        if isinstance(this_color, dict):
-            this_color = this_color[params['types'][inds[ci]]]
-        ax_vscroll.add_patch(mpl.patches.Rectangle((0, ci), 1, 1,
-                                                   facecolor=this_color,
-                                                   edgecolor=this_color))
+    n_ch = len(inds)
+
+    if 'fig_selection' in params:
+        selections = params['selections']
+        labels = [l._text for l in params['fig_selection'].radio.labels]
+        # Flatten the selections dict to a list.
+        cis = [item for sublist in [selections[l] for l in labels] for item
+               in sublist]
+
+        for idx, ci in enumerate(cis):
+            this_color = (bad_color if info['ch_names'][ci] in
+                          info['bads'] else color)
+            if isinstance(this_color, dict):
+                this_color = this_color[params['types'][ci]]
+            ax_vscroll.add_patch(mpl.patches.Rectangle((0, idx), 1, 1,
+                                                       facecolor=this_color,
+                                                       edgecolor=this_color))
+        ax_vscroll.set_ylim(len(cis), 0)
+        n_channels = max([len(selections[labels[0]]), n_channels])
+    else:
+        for ci in range(len(inds)):
+            this_color = (bad_color if info['ch_names'][inds[ci]] in
+                          info['bads'] else color)
+            if isinstance(this_color, dict):
+                this_color = this_color[params['types'][inds[ci]]]
+            ax_vscroll.add_patch(mpl.patches.Rectangle((0, ci), 1, 1,
+                                                       facecolor=this_color,
+                                                       edgecolor=this_color))
+        ax_vscroll.set_ylim(n_ch, 0)
     vsel_patch = mpl.patches.Rectangle((0, 0), 1, n_channels, alpha=0.5,
                                        facecolor='w', edgecolor='w')
     ax_vscroll.add_patch(vsel_patch)
     params['vsel_patch'] = vsel_patch
+
     hsel_patch = mpl.patches.Rectangle((params['t_start'], 0),
                                        params['duration'], 1, edgecolor='k',
                                        facecolor=(0.75, 0.75, 0.75),
@@ -601,18 +677,9 @@ def _prepare_mne_browse_raw(params, title, bgcolor, color, bad_color, inds,
     ax_hscroll.add_patch(hsel_patch)
     params['hsel_patch'] = hsel_patch
     ax_hscroll.set_xlim(0, params['n_times'] / float(info['sfreq']))
-    n_ch = len(info['ch_names'])
-    ax_vscroll.set_ylim(n_ch, 0)
+
     ax_vscroll.set_title('Ch.')
 
-    # make shells for plotting traces
-    _setup_browser_offsets(params, n_channels)
-    ax.set_xlim(params['t_start'], params['t_start'] + params['duration'],
-                False)
-
-    params['lines'] = [ax.plot([np.nan], antialiased=False, linewidth=0.5)[0]
-                       for _ in range(n_ch)]
-    ax.set_yticklabels(['X' * max([len(ch) for ch in info['ch_names']])])
     vertline_color = (0., 0.75, 0.)
     params['ax_vertline'] = ax.plot([0, 0], ax.get_ylim(),
                                     color=vertline_color, zorder=-1)[0]
@@ -622,13 +689,22 @@ def _prepare_mne_browse_raw(params, title, bgcolor, color, bad_color, inds,
     params['ax_hscroll_vertline'] = ax_hscroll.plot([0, 0], [0, 1],
                                                     color=vertline_color,
                                                     zorder=2)[0]
+    # make shells for plotting traces
+    _setup_browser_offsets(params, n_channels)
+    ax.set_xlim(params['t_start'], params['t_start'] + params['duration'],
+                False)
+
+    params['lines'] = [ax.plot([np.nan], antialiased=False, linewidth=0.5)[0]
+                       for _ in range(n_ch)]
+    ax.set_yticklabels(['X' * max([len(ch) for ch in info['ch_names']])])
 
 
-def _plot_raw_traces(params, inds, color, bad_color, event_lines=None,
+def _plot_raw_traces(params, color, bad_color, event_lines=None,
                      event_color=None):
     """Helper for plotting raw"""
     lines = params['lines']
     info = params['info']
+    inds = params['inds']
     n_channels = params['n_channels']
     params['bad_color'] = bad_color
     labels = params['ax'].yaxis.get_ticklabels()
@@ -640,7 +716,7 @@ def _plot_raw_traces(params, inds, color, bad_color, event_lines=None,
         # n_channels per view >= the number of traces available
         if ii >= len(lines):
             break
-        elif ch_ind < len(info['ch_names']):
+        elif ch_ind < len(inds):
             # scale to fit
             ch_name = info['ch_names'][inds[ch_ind]]
             tick_list += [ch_name]
@@ -722,7 +798,8 @@ def _plot_raw_traces(params, inds, color, bad_color, event_lines=None,
     params['ax'].set_xlim(params['times'][0],
                           params['times'][0] + params['duration'], False)
     params['ax'].set_yticklabels(tick_list)
-    params['vsel_patch'].set_y(params['ch_start'])
+    if 'fig_selection' not in params:
+        params['vsel_patch'].set_y(params['ch_start'])
     params['fig'].canvas.draw()
     # XXX This is a hack to make sure this figure gets drawn last
     # so that when matplotlib goes to calculate bounds we don't get a
@@ -815,3 +892,69 @@ def plot_raw_psd_topo(raw, tmin=0., tmax=None, fmin=0., fmax=100., proj=False,
     except TypeError:  # not all versions have this
         plt_show(show)
     return fig
+
+
+def _region_picked(event, params):
+    """Callback for selecting a region by clicking the topomap."""
+    ind = event.ind[0]
+    for type in ('mag', 'grad', 'eeg', 'seeg'):
+        if type in params['types']:
+            types = np.where(np.array(params['types']) == type)[0]
+            break
+    labels = [l._text for l in params['fig_selection'].radio.labels]
+    for idx, label in enumerate(labels):
+        if types[ind] in params['selections'][label]:
+            _set_radio_button(idx, params)
+            return
+
+
+def _setup_browser_selection(raw, kind):
+    """Helper for organizing browser selections."""
+    import matplotlib.pyplot as plt
+    from matplotlib.widgets import RadioButtons
+    from ..selection import (read_selection, _SELECTIONS, _EEG_SELECTIONS,
+                             _divide_to_regions)
+    from ..utils import _get_stim_channel
+    if kind == 'position':
+        order = _divide_to_regions(raw.info)
+        keys = _SELECTIONS[1:]  # no 'Vertex'
+    elif 'selection':
+        from ..io import RawFIF, RawArray
+        if not isinstance(raw, (RawFIF, RawArray)):
+            raise ValueError("order='selection' only works for Neuromag data. "
+                             "Use order='position' instead.")
+        order = dict()
+        try:
+            stim_ch = _get_stim_channel(None, raw.info)
+        except ValueError:
+            stim_ch = ['']
+        keys = np.concatenate([_SELECTIONS, _EEG_SELECTIONS])
+        stim_ch = pick_channels(raw.ch_names, stim_ch)
+        for key in keys:
+            channels = read_selection(key, info=raw.info)
+            picks = pick_channels(raw.ch_names, channels)
+            if len(picks) == 0:
+                continue  # omit empty selections
+            order[key] = np.concatenate([picks, stim_ch])
+
+    misc = pick_types(raw.info, meg=False, eeg=False, stim=True, eog=True,
+                      ecg=True, emg=True, ref_meg=False, misc=True, resp=True,
+                      chpi=True, exci=True, ias=True, syst=True, seeg=False,
+                      bio=True, ecog=False, exclude=())
+    if len(misc) > 0:
+        order['Misc'] = misc
+    keys = np.concatenate([keys, ['Misc']])
+    fig_selection = figure_nobar(figsize=(2, 6), dpi=80)
+    fig_selection.canvas.set_window_title('Selection')
+    rax = plt.subplot2grid((6, 1), (2, 0), rowspan=4, colspan=1)
+    topo_ax = plt.subplot2grid((6, 1), (0, 0), rowspan=2, colspan=1)
+    plot_sensors(raw.info, kind='topomap', ch_type=None, axes=topo_ax,
+                 ch_groups=kind, title='', show=False)
+    fig_selection.radio = RadioButtons(rax, [key for key in keys
+                                             if key in order.keys()])
+
+    for circle in fig_selection.radio.circles:
+        circle.set_radius(0.02)  # make them smaller to prevent overlap
+        circle.set_edgecolor('gray')  # make sure the buttons are visible
+
+    return order, fig_selection